crud acl system and check permissions

+++ app/Http/Controllers/ClientsController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\ClientRequest;
 use App\Http\Requests\UpdateClient;
 use App\Models\Client;
 use App\Models\User;
+use Illuminate\Support\Facades\Http;
 use Illuminate\Support\Facades\Request;
 use Illuminate\Support\Facades\URL;
 use Illuminate\Validation\Rule;
@@ -20,6 +21,10 @@ class ClientsController extends Controller
      */
     public function index()
     {
+        if (!auth()->user()->can('client.index')){
+            abort(404);
+        }
+
         return inertia('Modules/Clients/Index', [
             $search = Request::input('search'),
             'clients' => Client::query()->with('projects')
@@ -71,7 +76,9 @@ class ClientsController extends Controller
      */
     public function store(ClientRequest $request)
     {
-
+        if (!auth()->user()->can('client.create')) {
+            abort(401 );
+        }
        $data = $request->validate([
             "name" => ['required'],
             "email" => ['required', 'email', Rule::unique('clients', 'email')],
@@ -104,6 +111,9 @@ class ClientsController extends Controller
     public function show($id)
     {
 
+        if (!auth()->user()->can('client.show')) {
+            abort(401 );
+        }
         $user = Client::findOrFail($id)->load('users','transactions','transactions.user',
             'transactions.method','customeInvoices',
             'quotations','quotations.user', 'projects',
@@ -133,6 +143,9 @@ class ClientsController extends Controller
      */
     public function update(UpdateClient $request, Client $client)
     {
+        if (!auth()->user()->can('client.edit')) {
+            abort(401 );
+        }
         $client->update($request->validated());
         if ($request->agents){
             $client->users()->sync($request->input('agents'));
@@ -148,6 +161,9 @@ class ClientsController extends Controller
      */
     public function destroy($id)
     {
+        if (!auth()->user()->can('client.delete')) {
+            abort(401 );
+        }
         Client::findOrFail($id)->delete();
         return redirect()->route('clients.index');
     }
